Fix location of tests after all old tests were moved
